refactor(billing): split BillingPdfService into focused pdf/ modules
Pure move, no behavior change (verified byte-identical output across
deposit/withdraw/monthly-bill). The 929-line file became:

- pdf/types.ts          BillingPdfDocumentDto
- pdf/primitives.ts     PageCursor, page/color consts, sanitizeForWinAnsi,
                        formatMoney, fiatFor, VAT/fee rates, helpers
- pdf/copy.ts           FRAK_SELLER + French Objet/attestation strings
- pdf/shared.ts         brand logo, footer, header, party blocks, amounts,
                        attestation (plain functions, no inheritance)
- pdf/DepositDocument.ts   deposit body
- pdf/WithdrawDocument.ts  withdraw body
- pdf/MonthlyBillDocument.ts  bill header, reward table, groupRewards,
                        TVA/recap, ledger status
- pdf/index.ts          BillingPdfService.render() dispatch

BillingPdfService.ts is now a re-export shim so all importers and the
existing test resolve unchanged. 88 billing tests pass; tsc + biome clean.
